+++ rumqttc-v4/examples/tls.rs
@@ -25,7 +25,7 @@ async fn main() -> Result<(), Box<dyn Error>> {
 
     mqttoptions.set_transport(Transport::tls_with_config(client_config.into()));
 
-    let (_client, mut eventloop) = AsyncClient::builder(mqttoptions).capacity(10).build_async();
+    let (_client, mut eventloop) = AsyncClient::builder(mqttoptions).capacity(10).build();
 
     loop {
         match eventloop.poll().await {