+++ rumqttc-v4/tests/mqtt_options.rs
@@ -24,9 +24,7 @@ async fn v4_custom_socket_connector_is_invoked() {
 
     assert!(options.has_socket_connector());
 
-    let (_client, mut eventloop) = rumqttc::AsyncClient::builder(options)
-        .capacity(10)
-        .build_async();
+    let (_client, mut eventloop) = rumqttc::AsyncClient::builder(options).capacity(10).build();
     let error = eventloop
         .poll()
         .await